Add script to build Ardour on Mac OS
This downloads ardour-build-tools and ardour repos to a temporary workdir,
runs 'build-stack' to download, build and install all dependencies to a
path in the user's homedir, builds Ardour using the correct dependencies,
and package it as a DMG file.

At the end of the process it cleans-up all build artifacts, leaving only
the DMG file on the directory where the script was started from.